server/requirements.txt
attrs==22.2.0
autopep8==1.7.0
blinker==1.5
certifi==2022.12.7
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
dnspython==2.3.0
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
fonttools==4.34.4
idna==3.3
importlib-metadata==6.0.0
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jycm==1.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
numpy==1.23.1
packaging==21.3
pluggy==1.0.0
pycodestyle==2.9.1
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
requests==2.28.1
sentry-sdk==1.16.0
six==1.16.0
tabulate==0.8.10
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.3
zipp==3.15.0
